Monotone-Parent: f051bedf15bee8178f44e8de5fa444e19f546ceb Monotone-Revision: 68da0e360669d9370d0d64e6216124fa814549a9
Monotone-Author: wsourdeau@inverse.ca
Monotone-Date: 2010-08-30T19:42:03
Monotone-Branch: ca.inverse.sope

+++ sope-appserver/NGObjWeb/DynamicElements/_WOStaticHTMLElement.h
@@ -38,8 +38,8 @@
   NSString *text;
 }
 
-- (id)initWithBuffer:(const char *)_buffer length:(unsigned)_len;
-- (id)initWithCharacters:(const unichar *)_buffer length:(unsigned)_len;
+- (id)initWithBuffer:(const char *)_buffer length:(NSUInteger)_len;
+- (id)initWithCharacters:(const unichar *)_buffer length:(NSUInteger)_len;
 
 @end
