+++ sope-xml/SaxObjC/SaxAttributeList.m
@@ -107,13 +107,13 @@
 
 /* protocol implementation */
 
-- (NSString *)nameAtIndex:(unsigned int)_idx {
+- (NSString *)nameAtIndex:(NSUInteger)_idx {
   return [self->names objectAtIndex:_idx];
 }
-- (NSString *)typeAtIndex:(unsigned int)_idx {
+- (NSString *)typeAtIndex:(NSUInteger)_idx {
   return [self->types objectAtIndex:_idx];
 }
-- (NSString *)valueAtIndex:(unsigned int)_idx {
+- (NSString *)valueAtIndex:(NSUInteger)_idx {
   return [self->values objectAtIndex:_idx];
 }
 
@@ -134,7 +134,7 @@
   return [self valueAtIndex:i];
 }
 
-- (unsigned int)count {
+- (NSUInteger)count {
   return [self->names count];
 }
